Renamed lookout.sdk.grpc.utils to lookout.sdk.grpc.connection
